requirements.txt
requests==2.22.0
python-dotenv==0.10.3
terminaltables==3.1.0